backend/pyproject.toml
[project]
name = "personal_finance_tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosmtplib>=5.0.0",
    "bcrypt>=5.0.0",
    "cryptography>=46.0.3",
    "fastapi>=0.122.0",
    "google-generativeai>=0.8.5",
    "lightgbm>=4.6.0",
    "mlflow>=3.6.0",
    "mysql-connector-python>=9.5.0",
    "nltk>=3.9.2",
    "numpy>=2.3.5",
    "onnxruntime>=1.23.2",
    "openapi>=2.0.0",
    "pandas>=2.3.3",
    "passlib>=1.7.4",
    "pydantic>=2.12.5",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=7.1.0",
    "spacy>=3.8.11",
    "sqlalchemy>=2.0.44",
    "tensorflow>=2.20.0",
    "torch>=2.9.1",
    "transformers>=4.57.3",
    "uvicorn>=0.38.0",
    "xgboost>=3.1.2",
]
